chore: move GitHub types into own sub-folder
